Made zmqpp::message a template class
Renamed message to message_base and redefined message to `typedef
message_base<std::vector> message` to avoid breaking the current API.
This commit is to allow the ability of changing the container without
changing the base class. Currently, it only supports `std::vector` and
`std::deque`.
