+++ src/zmqpp/message.hpp
@@ -14,14 +14,17 @@
  * \author Ben Gray (\@benjamg)
  */
 
-#ifndef ZMQPP_MESSAGE_HPP_
-#define ZMQPP_MESSAGE_HPP_
+#ifndef ZMQPP_message_base_HPP_
+#define ZMQPP_message_base_HPP_
 
 #include <cassert>
 #include <functional>
 #include <string>
+#include <cstring>
+#include <cassert>
 #include <unordered_map>
 #include <vector>
+#include <deque>
 #include <utility>
 
 #include <zmq.h>
@@ -29,18 +32,20 @@
 #include "compatibility.hpp"
 #include "frame.hpp"
 #include "signal.hpp"
+#include "exception.hpp"
 
 namespace zmqpp
 {
 
 /**
- * \brief a zmq message with optional multipart support
+ * \brief a zmq message_base with optional multipart support
  *
- * A zmq message is made up of one or more parts which are sent together to
- * the target endpoints. zmq guarantees either the whole message or none
- * of the message will be delivered.
+ * A zmq message_base is made up of one or more parts which are sent together to
+ * the target endpoints. zmq guarantees either the whole message_base or none
+ * of the message_base will be delivered.
  */
-class ZMQPP_EXPORT message
+template<template<class T, class = std::allocator<T> > class container_type>
+class message_base
 {
 public:
 	/**
@@ -55,39 +60,137 @@ public:
 	 */
 	typedef std::function<void (void*)> release_function;
 
-	message();
-	~message();
+	message_base()
+		: _parts()
+		, _read_cursor(0)
+	{
+		static_assert(std::is_same<parts_type, std::vector<frame>>::value || std::is_same<parts_type, std::deque<frame>>::value, "Container not supported.");
+	}
+	~message_base()
+	{
+		clear();
+	}
 
     template <typename T, typename ...Args>
-    message(T const &part, Args &&...args)
-        : message()
+    message_base(T const &part, Args &&...args)
+        : message_base()
     {
         add(part, std::forward<Args>(args)...);
     }
 
-	size_t parts() const;
-	size_t size(size_t const part) const;
-	std::string get(size_t const part) const;
+	size_t parts() const
+	{
+		return _parts.size();
+	}
+	size_t size(size_t const part) const
+	{
+		if (part >= _parts.size())
+		{
+			throw exception("attempting to request a message_base part outside the valid range");
+		}
 
-	void get(int8_t& integer, size_t const part) const;
-	void get(int16_t& integer, size_t const part) const;
-	void get(int32_t& integer, size_t const part) const;
-	void get(int64_t& integer, size_t const part) const;	
-	void get(signal& sig, size_t const part) const;
+		return _parts[part].size();
+	}
+	std::string get(size_t const part) const
+	{
+		return std::string(static_cast<char const*>(raw_data(part)), size(part));
+	}
 
-	void get(uint8_t& unsigned_integer, size_t const part) const;
-	void get(uint16_t& unsigned_integer, size_t const part) const;
-	void get(uint32_t& unsigned_integer, size_t const part) const;
-	void get(uint64_t& unsigned_integer, size_t const part) const;
+	void get(int8_t& integer, size_t const part) const
+	{
+		assert(sizeof(int8_t) == size(part));
 
-	void get(float& floating_point, size_t const part) const;
-	void get(double& double_precision, size_t const part) const;
-	void get(bool& boolean, size_t const part) const;
+		int8_t const* byte = static_cast<int8_t const*>(raw_data(part));
+		integer = *byte;
+	}
+	void get(int16_t& integer, size_t const part) const
+	{
+		assert(sizeof(int16_t) == size(part));
 
-	void get(std::string& string, size_t const part) const;
+		uint16_t const* network_order = static_cast<uint16_t const*>(raw_data(part));
+		integer = static_cast<int16_t>(ntohs(*network_order));
+	}
+	void get(int32_t& integer, size_t const part) const
+	{
+		assert(sizeof(int32_t) == size(part));
 
-	// Warn: If a pointer type is requested the message (well zmq) still 'owns'
-	// the data and will release it when the message object is freed.
+		uint32_t const* network_order = static_cast<uint32_t const*>(raw_data(part));
+		integer = static_cast<int32_t>(htonl(*network_order));
+	}
+	void get(int64_t& integer, size_t const part) const
+	{
+		assert(sizeof(int64_t) == size(part));
+
+		uint64_t const* network_order = static_cast<uint64_t const*>(raw_data(part));
+		integer = static_cast<int64_t>(htonll(*network_order));
+	}
+	void get(signal& sig, size_t const part) const
+	{
+		assert(sizeof(signal) == size(part));
+		int64_t v;
+		get(v, part);
+
+		sig = static_cast<signal>(v);
+	}
+
+	void get(uint8_t& unsigned_integer, size_t const part) const
+	{
+		assert(sizeof(uint8_t) == size(part));
+
+		uint8_t const* byte = static_cast<uint8_t const*>(raw_data(part));
+		unsigned_integer = *byte;
+	}
+	void get(uint16_t& unsigned_integer, size_t const part) const
+	{
+		assert(sizeof(uint16_t) == size(part));
+
+		uint16_t const* network_order = static_cast<uint16_t const*>(raw_data(part));
+		unsigned_integer = ntohs(*network_order);
+	}
+	void get(uint32_t& unsigned_integer, size_t const part) const
+	{
+		assert(sizeof(uint32_t) == size(part));
+
+		uint32_t const* network_order = static_cast<uint32_t const*>(raw_data(part));
+		unsigned_integer = ntohl(*network_order);
+	}
+	void get(uint64_t& unsigned_integer, size_t const part) const
+	{
+		assert(sizeof(uint64_t) == size(part));
+
+		uint64_t const* network_order = static_cast<uint64_t const*>(raw_data(part));
+		unsigned_integer = ntohll(*network_order);
+	}
+
+	void get(float& floating_point, size_t const part) const
+	{
+		assert(sizeof(float) == size(part));
+
+		float const* network_order = static_cast<float const*>(raw_data(part));
+		floating_point = zmqpp::ntohf(*network_order);
+	}
+	void get(double& double_precision, size_t const part) const
+	{
+		assert(sizeof(double) == size(part));
+
+		double const* network_order = static_cast<double const*>(raw_data(part));
+		double_precision = zmqpp::ntohd(*network_order);
+	}
+	void get(bool& boolean, size_t const part) const
+	{
+		assert(sizeof(uint8_t) == size(part));
+
+		uint8_t const* byte = static_cast<uint8_t const*>(raw_data(part));
+		boolean = (*byte != 0);
+	}
+
+	void get(std::string& string, size_t const part) const
+	{
+		string.assign(get(part));
+	}
+
+	// Warn: If a pointer type is requested the message_base (well zmq) still 'owns'
+	// the data and will release it when the message_base object is freed.
 	template<typename Type>
 	Type get(size_t const part)
 	{
@@ -112,24 +215,30 @@ public:
     }
 
 	// Raw get data operations, useful with data structures more than anything else
-	// Warn: The message (well zmq) still 'owns' the data and will release it
-	// when the message object is freed.
+	// Warn: The message_base (well zmq) still 'owns' the data and will release it
+	// when the message_base object is freed.
 	template<typename Type>
 	void get(Type*& value, size_t const part) const
 	{
 		value = static_cast<Type*>(raw_data(part));
 	}
 
-	// Warn: The message (well zmq) still 'owns' the data and will release it
-	// when the message object is freed.
+	// Warn: The message_base (well zmq) still 'owns' the data and will release it
+	// when the message_base object is freed.
 	template<typename Type>
 	void get(Type** value, size_t const part) const
 	{
 		*value = static_cast<Type*>(raw_data(part));
 	}
 
-	// Move operators will take ownership of message parts without copying
-	void move(void* part, size_t const size, release_function const& release);
+	// Move operators will take ownership of message_base parts without copying
+	void move(void* part, size_t const size, release_function const& release)
+	{
+		callback_releaser* hint = new callback_releaser();
+		hint->func = release;
+
+		_parts.push_back(frame(part, size, &message_base::release_callback, hint));
+	}
 
 	// Raw move data operation, useful with data structures more than anything else
 	template<typename Object>
@@ -161,7 +270,7 @@ public:
 
 	// Use exact data past, neither zmqpp nor 0mq will copy, alter or delete
 	// this data. It must remain as valid for at least the lifetime of the
-	// 0mq message, recommended only with const data.
+	// 0mq message_base, recommended only with const data.
 	template<typename Type>
 	ZMQPP_DEPRECATED("Use add_nocopy() or add_nocopy_const() instead.")
 	void add_const(Type *part, size_t const data_size)
@@ -177,7 +286,7 @@ public:
 	 * the underlying zmq_msg_t. Note that you cannot always know about
 	 * this lifetime, so be careful.
 	 *
-	 * @param part The pointed-to data that will be send in the message.
+	 * @param part The pointed-to data that will be send in the message_base.
 	 * @param data_size The number of byte pointed-to by "part".
 	 * @param ffn The free function called by libzmq when it doesn't need
 	 * your buffer anymore. It defaults to nullptr, meaning your data
@@ -217,7 +326,7 @@ public:
 	 * will happily return a non-const pointer to your data. It's your responsibility
 	 * to not modify it.
 	 *
-	 * @param part The pointed-to data that will be send in the message.
+	 * @param part The pointed-to data that will be send in the message_base.
 	 * @param data_size The number of byte pointed-to by "part".
 	 * @param ffn The free function called by libzmq when it doesn't need
 	 * your buffer anymore. It defaults to nullptr, meaning your data
@@ -238,57 +347,212 @@ public:
 	}
 
 	// Stream reader style
-	void reset_read_cursor();
+	void reset_read_cursor()
+	{
+		_read_cursor = 0;
+	}
 
 	template<typename Type>
-	message& operator>>(Type& value)
+	message_base& operator>>(Type& value)
 	{
 		get(value, _read_cursor++);
 		return *this;
 	}
 
 	// Stream writer style - these all use copy styles
-	message& operator<<(int8_t const integer);
-	message& operator<<(int16_t const integer);
-	message& operator<<(int32_t const integer);
-	message& operator<<(int64_t const integer);
-	message& operator<<(signal const sig);
+	message_base& operator<<(int8_t const integer)
+	{
+		add_raw(reinterpret_cast<void const*>(&integer), sizeof(int8_t));
+		return *this;
+	}
+	message_base& operator<<(int16_t const integer)
+	{
+		uint16_t network_order = htons(static_cast<uint16_t>(integer));
+		add_raw(reinterpret_cast<void const*>(&network_order), sizeof(uint16_t));
 
-	message& operator<<(uint8_t const unsigned_integer);
-	message& operator<<(uint16_t const unsigned_integer);
-	message& operator<<(uint32_t const unsigned_integer);
-	message& operator<<(uint64_t const unsigned_integer);
+		return *this;
+	}
+	message_base& operator<<(int32_t const integer)
+	{
+		uint32_t network_order = htonl(static_cast<uint32_t>(integer));
+		add_raw(reinterpret_cast<void const*>(&network_order), sizeof(uint32_t));
 
-	message& operator<<(float const floating_point);
-	message& operator<<(double const double_precision);
-	message& operator<<(bool const boolean);
+		return *this;
+	}
+	message_base& operator<<(int64_t const integer)
+	{
+		uint64_t network_order = htonll(static_cast<uint64_t>(integer));
+		add_raw(reinterpret_cast<void const*>(&network_order), sizeof(uint64_t));
 
-	message& operator<<(char const* c_string);
-	message& operator<<(std::string const& string);
+		return *this;
+	}
+	message_base& operator<<(signal const sig)
+	{
+		return (*this) << static_cast<int64_t>(sig);
+	}
+
+	message_base& operator<<(uint8_t const unsigned_integer)
+	{
+		add_raw(reinterpret_cast<void const*>(&unsigned_integer), sizeof(uint8_t));
+		return *this;
+	}
+	message_base& operator<<(uint16_t const unsigned_integer)
+	{
+		uint16_t network_order = htons(unsigned_integer);
+		add_raw(reinterpret_cast<void const*>(&network_order), sizeof(uint16_t));
+
+		return *this;
+	}
+	message_base& operator<<(uint32_t const unsigned_integer)
+	{
+		uint32_t network_order = htonl(unsigned_integer);
+		add_raw(reinterpret_cast<void const*>(&network_order), sizeof(uint32_t));
+
+		return *this;
+	}
+	message_base& operator<<(uint64_t const unsigned_integer)
+	{
+		uint64_t network_order = htonll(unsigned_integer);
+		add_raw(reinterpret_cast<void const*>(&network_order), sizeof(uint64_t));
+
+		return *this;
+	}
+
+	message_base& operator<<(float const floating_point)
+	{
+		assert(sizeof(float) == 4);
+
+		float network_order = zmqpp::htonf(floating_point);
+		add_raw(&network_order, sizeof(float));
+
+		return *this;
+	}
+	message_base& operator<<(double const double_precision)
+	{
+		assert(sizeof(double) == 8);
+
+		double network_order = zmqpp::htond(double_precision);
+		add_raw(&network_order, sizeof(double));
+
+		return *this;
+	}
+	message_base& operator<<(bool const boolean)
+	{
+		uint8_t byte = (boolean) ? 1 : 0;
+		add_raw(reinterpret_cast<void const*>(&byte), sizeof(uint8_t));
+
+		return *this;
+	}
+
+	message_base& operator<<(char const* c_string)
+	{
+		add_raw(reinterpret_cast<void const*>(c_string), strlen(c_string));
+		return *this;
+	}
+	message_base& operator<<(std::string const& string)
+	{
+		add_raw(reinterpret_cast<void const*>(string.data()), string.size());
+		return *this;
+	}
 
 	// Queue manipulation
-	void push_front(void const* part, size_t const size);
+	void push_front(void const* part, size_t const size)
+	{
+		if (std::is_same<parts_type, std::vector<frame>>::value)
+		{
+			_parts.emplace(_parts.begin(), part, size);
+		}
+		else if (std::is_same<parts_type, std::deque<frame>>::value)
+		{
+			_parts.emplace_front(part, size);
+		}
+	}
 
 	// TODO: unify conversion of types with the stream operators
-	void push_front(int8_t const integer);
-	void push_front(int16_t const integer);
-	void push_front(int32_t const integer);
-	void push_front(int64_t const integer);
-	void push_front(signal const sig);
+	void push_front(int8_t const integer)
+	{
+		push_front(&integer, sizeof(int8_t));
+	}
+	void push_front(int16_t const integer)
+	{
+		uint16_t network_order = htons(static_cast<uint16_t>(integer));
+		push_front(&network_order, sizeof(uint16_t));
+	}
+	void push_front(int32_t const integer)
+	{
+		uint32_t network_order = htonl(static_cast<uint32_t>(integer));
+		push_front(&network_order, sizeof(uint32_t));
+	}
+	void push_front(int64_t const integer)
+	{
+		uint64_t network_order = htonll(static_cast<uint64_t>(integer));
+		push_front(&network_order, sizeof(uint64_t));
+	}
+	void push_front(signal const sig)
+	{
+		push_front(static_cast<int64_t>(sig));
+	}
 
-	void push_front(uint8_t const unsigned_integer);
-	void push_front(uint16_t const unsigned_integer);
-	void push_front(uint32_t const unsigned_integer);
-	void push_front(uint64_t const unsigned_integer);
+	void push_front(uint8_t const unsigned_integer)
+	{
+		push_front(&unsigned_integer, sizeof(uint8_t));
+	}
+	void push_front(uint16_t const unsigned_integer)
+	{
+		uint16_t network_order = htons(unsigned_integer);
+		push_front(&network_order, sizeof(uint16_t));
+	}
+	void push_front(uint32_t const unsigned_integer) 
+	{
+		uint32_t network_order = htonl(unsigned_integer);
+		push_front(&network_order, sizeof(uint32_t));
+	}
+	void push_front(uint64_t const unsigned_integer)
+	{
+		uint64_t network_order = htonll(unsigned_integer);
+		push_front(&network_order, sizeof(uint64_t));
+	}
 
-	void push_front(float const floating_point);
-	void push_front(double const double_precision);
-	void push_front(bool const boolean);
+	void push_front(float const floating_point)
+	{
+		assert(sizeof(float) == 4);
 
-	void push_front(char const* c_string);
-	void push_front(std::string const& string);
+		float network_order = zmqpp::htonf(floating_point);
+		push_front(&network_order, sizeof(float));
+	}
+	void push_front(double const double_precision)
+	{
+		assert(sizeof(double) == 8);
 
-	void pop_front();
+		double network_order = zmqpp::htond(double_precision);
+		push_front(&network_order, sizeof(double));
+	}
+	void push_front(bool const boolean)
+	{
+		uint8_t byte = (boolean) ? 1 : 0;
+		push_front(&byte, sizeof(uint8_t));
+	}
+
+	void push_front(char const* c_string)
+	{
+		push_front(c_string, strlen(c_string));
+	}
+	void push_front(std::string const& string)
+	{
+		push_front(string.data(), string.size());
+	}
+
+	void pop_front()
+	{
+		if (std::is_same<parts_type, std::vector<frame>>::value)
+		{
+			_parts.erase(_parts.begin());
+		}
+		else if (std::is_same<parts_type, std::deque<frame>>::value)
+		{
+			_parts.pop_front();
+		}
+	}
 
 	void push_back(void const* part, size_t const data_size)
 	{
@@ -301,34 +565,109 @@ public:
 		*this << part;
 	}
 
-	void pop_back();
+	void pop_back()
+	{
+		_parts.pop_back();
+	}
 
-	void remove(size_t const part);
+	void remove(size_t const part)
+	{
+		_parts.erase(_parts.begin() + part);
+	}
 
 	// Move supporting
-	message(message&& source) NOEXCEPT;
-	message& operator=(message&& source) NOEXCEPT;
+	message_base(message_base&& source) NOEXCEPT
+		: _parts()
+		, _read_cursor(source._read_cursor)
+	{
+		std::swap(_parts, source._parts);
+		source._read_cursor = 0;
+	}
+	message_base& operator=(message_base<container_type>&& source) NOEXCEPT
+	{
+		_read_cursor = source._read_cursor;
+		source._read_cursor = 0;
+		_parts = std::move(source._parts);
+		return *this;
+	}
 
 	// Copy support
-	message copy() const;
-	void copy(message const& source);
+	message_base copy() const
+	{
+		message_base<container_type> msg;
+		msg.copy(*this);
+		return msg;
+	}
+	void copy(message_base<container_type> const& source)
+	{
+		_parts.resize(source._parts.size());
+		for (size_t i = 0; i < source._parts.size(); ++i)
+		{
+			_parts[i] = source._parts[i].copy();
+		}
+
+		// we don't need a copy of the releasers as we did data copies of the internal data,
+		//_releasers = source._releasers;
+		//_strings = source._strings
+	}
 
 	// Used for internal tracking
-	void sent(size_t const part);
+	void sent(size_t const part)
+	{
+		// sanity check
+		assert(!_parts[part].is_sent());
+		_parts[part].mark_sent();
+	}
 
 	// Access to raw zmq details
-	void const* raw_data(size_t const part = 0) const;
-	zmq_msg_t& raw_msg(size_t const part = 0);
-	zmq_msg_t& raw_new_msg();
-	zmq_msg_t& raw_new_msg(size_t const reserve_data_size);
+	void const* raw_data(size_t const part = 0) const
+	{
+		if (part >= _parts.size())
+		{
+			throw zmqpp::exception("attempting to request a message_base part outside the valid range");
+		}
+
+		return _parts[part].data();
+	}
+	zmq_msg_t& raw_msg(size_t const part = 0)
+	{
+		if (part >= _parts.size())
+		{
+			throw zmqpp::exception("attempting to request a message_base part outside the valid range");
+		}
+
+		return _parts[part].msg();
+	}
+	zmq_msg_t& raw_new_msg()
+	{
+		_parts.push_back(frame());
+
+		return _parts.back().msg();
+	}
+	zmq_msg_t& raw_new_msg(size_t const reserve_data_size)
+	{
+		_parts.push_back(frame(reserve_data_size));
+
+		return _parts.back().msg();
+	}
 	
 	/**
-	 * Check if the message is a signal.
-	 * If the message has 1 part, has the correct size and if the 7 first bytes match
-	 * the signal header we consider the message a signal.
-	 * @return true if the message is a signal, false otherwise
+	 * Check if the message_base is a signal.
+	 * If the message_base has 1 part, has the correct size and if the 7 first bytes match
+	 * the signal header we consider the message_base a signal.
+	 * @return true if the message_base is a signal, false otherwise
 	 */
-	bool is_signal() const;
+	bool is_signal() const
+	{
+		if (parts() == 1 && size(0) == sizeof(signal))
+		{
+			signal s;
+			get(s, 0);
+			if ((static_cast<int64_t>(s) >> 8) == static_cast<int64_t>(signal::header))
+				return true;
+		}
+		return false;
+	}
 
 	/**
 	 * Gets the read cursor. For using get_raw() with stream-style reading.
@@ -336,7 +675,7 @@ public:
 	size_t read_cursor() const NOEXCEPT { return _read_cursor; }
 
 	/**
-	 * Gets the remaining number of parts in the message.
+	 * Gets the remaining number of parts in the message_base.
 	 */
 	size_t remaining() const NOEXCEPT { return  _parts.size() - _read_cursor; }
 
@@ -346,27 +685,57 @@ public:
 	 */
 	size_t next() NOEXCEPT { return ++_read_cursor; }
 
+	void clear() { _parts.clear(); }
+
 
 #if (ZMQ_VERSION_MAJOR == 4 && ZMQ_VERSION_MINOR >= 1)
 	/**
-	* Attemps to retrieve a metadata property from a message.
+	* Attemps to retrieve a metadata property from a message_base.
 	* The underlying call is `zmq_msg_gets()`.
 	*
-	* @note The message MUST have at least one frame, otherwise this wont work.
+	* @note The message_base MUST have at least one frame, otherwise this wont work.
 	*/
-	bool get_property(const std::string &property, std::string &out);
+	bool get_property(const std::string &property, std::string &out)
+	{
+		zmq_msg_t *zmq_raw_msg;
+		try
+		{
+			zmq_raw_msg = &raw_msg();
+		}
+		catch (zmqpp::exception const&) // empty
+		{
+			return false;
+		}
+
+		const char *property_value = zmq_msg_gets(zmq_raw_msg, property.c_str());
+		if (property_value == NULL)
+		{
+			// EINVAL is the only error code
+			assert(errno == EINVAL);
+			return false;
+		}
+
+		out = std::string(property_value);
+		return true;
+	}
 #endif
 
 private:
-	typedef std::vector<frame> parts_type;
+	typedef container_type<frame> parts_type;
 	parts_type _parts;
 	size_t _read_cursor;
 
 	// Disable implicit copy support, code must request a copy to clone
-	message(message const&) NOEXCEPT ZMQPP_EXPLICITLY_DELETED;
-	message& operator=(message const&) NOEXCEPT ZMQPP_EXPLICITLY_DELETED;
+	message_base(message_base const&) NOEXCEPT ZMQPP_EXPLICITLY_DELETED;
+	message_base& operator=(message_base const&) NOEXCEPT ZMQPP_EXPLICITLY_DELETED;
 
-	static void release_callback(void* data, void* hint);
+	static void release_callback(void* data, void* hint)
+	{
+		callback_releaser* releaser = static_cast<callback_releaser*>(hint);
+		releaser->func(data);
+
+		delete releaser;
+	}
 
 	template<typename Object>
 	static void deleter_callback(void* data)
@@ -375,6 +744,8 @@ private:
 	}
 };
 
+typedef message_base<std::vector> message;
+
 }
 
-#endif /* ZMQPP_MESSAGE_HPP_ */
+#endif /* ZMQPP_message_base_HPP_ */
