+++ src/zmqpp/socket.hpp
@@ -18,6 +18,7 @@
 #define ZMQPP_SOCKET_HPP_
 
 #include <cstring>
+#include <cassert>
 #include <string>
 #include <list>
 
@@ -34,11 +35,10 @@ namespace zmqpp
 {
 
 class context;
-class message;
+template<template<class T, class = std::allocator<T> > class container_type> class message_base;
 
 typedef std::string endpoint_t;
 typedef context     context_t;
-typedef message     message_t;
 
 #if (ZMQ_VERSION_MAJOR >= 4)
 namespace event
@@ -202,7 +202,67 @@ public:
 	 * \param dont_block boolean to dictate if we wait while sending.
 	 * \return true if message sent, false if it would have blocked
 	 */
-	bool send(message_t& message, bool const dont_block = false);
+	template<template<class T, class = std::allocator<T>> class container>
+	bool send(message_base<container>& message, bool const dont_block = false)
+	{
+		size_t parts = message.parts();
+		if (parts == 0)
+		{
+			throw std::invalid_argument("sending requires messages have at least one part");
+		}
+
+		bool dont_wait = dont_block;
+		for (size_t i = 0; i < parts; ++i)
+		{
+			int flag = socket::normal;
+			if (dont_wait) { flag |= socket::dont_wait; }
+			if (i < (parts - 1)) { flag |= socket::send_more; }
+
+#if (ZMQ_VERSION_MAJOR == 2)
+			int result = zmq_send(_socket, &message.raw_msg(i), flag);
+#elif (ZMQ_VERSION_MAJOR < 3) || ((ZMQ_VERSION_MAJOR == 3) && (ZMQ_VERSION_MINOR < 2))
+			int result = zmq_sendmsg(_socket, &message.raw_msg(i), flag);
+#else
+			int result = zmq_msg_send(&message.raw_msg(i), _socket, flag);
+#endif
+
+			if (result < 0)
+			{
+				// the zmq framework should not block if the first part is accepted
+				// so we should only ever get this error on the first part
+				if ((0 == i) && (EAGAIN == zmq_errno()))
+				{
+					return false;
+				}
+
+				if (EINTR == zmq_errno())
+				{
+					if (0 == i) // If first part of the message.
+					{
+						return false;
+					}
+
+					// If we have an interrupt but it's not on the first part then we
+					// know we can safely send out the rest of the message as we can
+					// enforce that it won't wait on a blocking action
+					dont_wait = true;
+					continue;
+				}
+
+				// sanity checking
+				assert(EAGAIN != zmq_errno());
+
+				throw zmq_internal_exception();
+			}
+
+			message.sent(i);
+		}
+
+		// Leave message reference in a stable state
+		zmqpp::message_base<container> local;
+		std::swap(local, message);
+		return true;
+	}
 
 	/**
 	 * Gets a message from the connection, this may be a multipart message.
@@ -214,7 +274,62 @@ public:
 	 * \param dont_block boolean to dictate if we wait for data.
 	 * \return true if message sent, false if it would have blocked
 	 */
-	bool receive(message_t& message, bool const dont_block = false);
+	template<template<class T, class = std::allocator<T>> class container>
+	bool receive(message_base<container>& message, bool const dont_block = false)
+	{
+		if (message.parts() > 0)
+		{
+			// swap and discard old message
+			zmqpp::message_base<container> local;
+			std::swap(local, message);
+		}
+
+		int flags = (dont_block) ? socket::dont_wait : socket::normal;
+		bool more = true;
+
+		while (more)
+		{
+#if (ZMQ_VERSION_MAJOR == 2)
+			int result = zmq_recv(_socket, &_recv_buffer, flags);
+#elif (ZMQ_VERSION_MAJOR < 3) || ((ZMQ_VERSION_MAJOR == 3) && (ZMQ_VERSION_MINOR < 2))
+			int result = zmq_recvmsg(_socket, &_recv_buffer, flags);
+#else
+			int result = zmq_msg_recv(&_recv_buffer, _socket, flags);
+#endif
+
+			if (result < 0)
+			{
+				if ((0 == message.parts()) && (EAGAIN == zmq_errno()))
+				{
+					return false;
+				}
+
+				if (EINTR == zmq_errno())
+				{
+					if (0 == message.parts())
+					{
+						return false;
+					}
+
+					// If we have an interrupt but it's not on the first part then we
+					// know we can safely pull out the rest of the message as it will
+					// not be blocking
+					continue;
+				}
+
+				assert(EAGAIN != zmq_errno());
+
+				throw zmq_internal_exception();
+			}
+
+			zmq_msg_t& dest = message.raw_new_msg();
+			zmq_msg_move(&dest, &_recv_buffer);
+
+			get(socket_option::receive_more, more);
+		}
+
+		return true;
+	}
 
 	/**
 	 * Sends the byte data held by the string as the next message part.
@@ -574,7 +689,8 @@ private:
 	socket(socket const&) NOEXCEPT ZMQPP_EXPLICITLY_DELETED;
 	socket& operator=(socket const&) NOEXCEPT ZMQPP_EXPLICITLY_DELETED;
 
-	void track_message(message_t const&, uint32_t const, bool&);
+	template<template<class T, class = std::allocator<T>> class container>
+	void track_message(message_base<container> const&, uint32_t const, bool&);
 };
 
 }
